- 회원가입, 업데이트 : 면접자 부분 수정
- 게시판 댓글 추가(기능구현 및 css)

+++ khis/src/main/java/com/kh/khis/login_join_mypage/model/vo/Member.java
@@ -36,7 +36,7 @@ public class Member implements Serializable {
 	private String kind;
 	private Date joinDate;
 	private String id;
-	private Integer interviewerNo; //Integer null을 쓸수있다.
+	private Long interviewerNo; //Integer null을 쓸수있다.
 	private Integer memberInfoNo;
 
 	/*
